tests: Extract osmorphing_utils from integration test utils
Move OS morphing helpers (write_os_image_to_disk, make_luks_device,
path_exists_on_device, etc.) into a dedicated osmorphing_utils module.

Refactored utils to use a context manager, simplifying the code.
